+++ cli/commands/metadata_export_test.go
@@ -57,15 +57,22 @@ var testMetadataFileMode = func(projectDirectory string) {
 		Eventually(isYAML(fileContents)).Should(BeTrue())
 	})
 }
+var verifyRequestTransformsMetadataIsExported = func(projectDirectory string) {
+	actionsMetadata := filepath.Join(projectDirectory, "metadata", "actions.yaml")
+	Expect(actionsMetadata).To(BeAnExistingFile())
+	b, err := ioutil.ReadFile(actionsMetadata)
+	Expect(err).To(BeNil())
+	Expect(string(b)).To(ContainSubstring("request_transform"))
+}
 
 var _ = Describe("hasura metadata export (config v3)", func() {
-	var projectDirectory string
+	var projectDirectory, hgeEndpoint string
 	var teardown func()
 	sourceName := randomdata.SillyName()
 	BeforeEach(func() {
 		projectDirectory = testutil.RandDirName()
 		hgeEndPort, teardownHGE := testutil.StartHasuraWithMetadataDatabase(GinkgoT(), testutil.HasuraDockerImage)
-		hgeEndpoint := fmt.Sprintf("http://0.0.0.0:%s", hgeEndPort)
+		hgeEndpoint = fmt.Sprintf("http://0.0.0.0:%s", hgeEndPort)
 
 		connectionString, teardownPG := AddDatabaseToHasura(hgeEndpoint, sourceName, "postgres")
 
@@ -98,6 +105,7 @@ var _ = Describe("hasura metadata export (config v3)", func() {
 			Eventually(session, timeout).Should(Exit(0))
 			Expect(session.Err.Contents()).Should(ContainSubstring("Metadata exported"))
 			Expect(filepath.Join(projectDirectory, "metadata", "databases", sourceName, "tables", "public_albums.yaml")).Should(BeAnExistingFile())
+			verifyRequestTransformsMetadataIsApplied(hgeEndpoint)
 		})
 		testExportMetadataToStdout(projectDirectory)
 		testMetadataFileMode(projectDirectory)
@@ -106,23 +114,26 @@ var _ = Describe("hasura metadata export (config v3)", func() {
 })
 
 var _ = Describe("hasura metadata export (config v2)", func() {
-	var projectDirectory string
+	var projectDirectory, hgeEndpoint string
 	var teardown func()
 	BeforeEach(func() {
 		projectDirectory = testutil.RandDirName()
 		hgeEndPort, teardownHGE := testutil.StartHasura(GinkgoT(), testutil.HasuraDockerImage)
-		hgeEndpoint := fmt.Sprintf("http://0.0.0.0:%s", hgeEndPort)
+		hgeEndpoint = fmt.Sprintf("http://0.0.0.0:%s", hgeEndPort)
 
 		// clone template project directory as test project directory
 		copyTestConfigV2Project(projectDirectory)
 		editEndpointInConfig(filepath.Join(projectDirectory, defaultConfigFilename), hgeEndpoint)
 		testutil.RunCommandAndSucceed(testutil.CmdOpts{
-			Args:             []string{"migrate", "apply", "&&", testutil.CLIBinaryPath, "metadata", "apply"},
+			Args:             []string{"migrate", "apply"},
+			WorkingDirectory: projectDirectory,
+		})
+		testutil.RunCommandAndSucceed(testutil.CmdOpts{
+			Args:             []string{"metadata", "apply"},
 			WorkingDirectory: projectDirectory,
 		})
 		// remove the directory after apply to check if it gets recreated after a successful export
 		Expect(os.RemoveAll(filepath.Join(projectDirectory, "metadata"))).To(BeNil())
-
 		teardown = func() {
 			os.RemoveAll(projectDirectory)
 			teardownHGE()
@@ -140,6 +151,8 @@ var _ = Describe("hasura metadata export (config v2)", func() {
 			Eventually(session, timeout).Should(Exit(0))
 			Expect(session.Err.Contents()).Should(ContainSubstring("Metadata exported"))
 			Expect(filepath.Join(projectDirectory, "metadata", "tables.yaml")).Should(BeAnExistingFile())
+
+			verifyRequestTransformsMetadataIsApplied(hgeEndpoint)
 		})
 		testExportMetadataToStdout(projectDirectory)
 		testMetadataFileMode(projectDirectory)