+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - console: allow dropping indices on all schemas
 - console: fix bug with displaying 1-to-1 relationship with the same column mapping (close #7552)
 - cli: split remote schema permissions metadata into seperate files (#7033)
+- cli: support action request transforms in metadata
 
 
 ### Function field names customization (#7405)